Yipe! Exclude test/asset files from publishing
I accidentally published a 22MB crate, not realizing that all the repo
files would get uploaded. I've asked crates.io what to do (I can yank
but not delete), and I'm now learning about the `exclude` directive in
Cargo.toml.

This fix is going to break the widgets and the examples, but I'd rather
err on the side of too little than keep uploading too much.